+++ RELEASENOTES.md
@@ -9,7 +9,7 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
 * Bugfix - UI updates in UltimateToolbar component
 * Bugfix - Registration table UI fixes
 * Bugfix - App Release Process - Fixed browser back button issue
-* Bugfix - App Release Process - Fixed user directing to bottom of the page
+* Bugfix - App Release Process - Fixed user directing to bottom of the page, changed notifications
 
 ### 0.5.4